Redesigned how kun'yomi and on'yomi are stored
I decided that storing the kun'yomi and on'yomi as elements was really a bit too much faff. It takes up too much space, and separating them as individual pieces is really unnecessary, since not only can I use a delimiter, but in reference materials I'll need to display all the kun'yomi and on'yomi together, and stitching them together is slightly more of a faff than separating them, if we're honest. So I decided to place them as attributes instead. This is essentially a reversion of a really early commit, when I set them up as elements in the first place (2849e9769bdb2086371de20fb33754219619c84a). I also set the on'yomi as a required attribute because it seems that while not every kanji has a kun'yomi, they all have at least one on'yomi. Also, I could not find a way to force at least one of them to be required without specifying which one, like I did when they were elements. If I find a kanji that has kun'yomi but no on'yomi I'll really be in trouble.