Redesigned refreshStaticQuestionBank method
In a partial reversion of an earlier commit (2f69ea49a0d13a129efb256a296be863ad8574a3), I redesigned the refreshStaticQuestionBank method to return nothing, and instead automatically run the newQuestion method to call up the first question. I also reorganized the related code in MainQuiz, so it no longer needs a nextQuestion method. Instead, the call to the newQuestion method from the QuestionBank class is in the only place it needs to be, right after an answer is submitted. And instead of the delay handler calling the nextQuestion method, it calls the refreshDisplay method. This makes things a bit simpler.