Moved fetch score call
Moved the code that refreshes the score display into the onCreate method. Because it doesn't need to run every time one returns from an options screen. In fact, it shouldn't change during that time either, no matter what. It should only change at the end of the day. But if it doesn't, that's really no big deal. I'd rather just keep the score building until the entire app reloads.