Simplified code for setting lblResponse colour
As part of my attempts to fix an issue involving text being set to the wrong colour (#7), I edited the code that sets the colour on lblResponse. I don't know if it'll make a lick of difference, but I just removed the getApplicationContext call, because I figure I should just be doing it from the current context. It's an unnecessary call that appears to be doing nothing but causing problems.