Removed getApplicationContext
All the initialization methods used the getApplicationContext method to ensure we were getting the application-level object, and not any other. I realize now that this is not necessary, and just results in cycles wasted. So I'm stripping them out, and redesigning the initialization methods to require the Application objects instead. Seems like a much more efficient way of doing things.