Added kana questions to locale refresh
The hook I added to reload the vocabulary questions on locale change has been working fine (5d34b13781db1e6d9fbee5e6813b6a11a0e3fb10). If the locale changes the vocabulary questions change accordingly so the entire app is running on the same language. However I did miss one thing. Seems the titles for the sets of hiragana and katakana questions are still in the previous language. I'm guessing because those items are stored as basic strings, and not resource references, they don't get updated automatically when the locale changes. So I have to deal with that in the same hook I placed for the vocabulary questions.