Moved locale change hook
The code to refresh the questions on a locale change has been moved to the KanaQuizMain class, so it's only established once, when the app starts. This is so I can simplify the code within, so it only calls the QuestionManagement initialize method, because before it was just running the exact same initialize code, which meant we had a bit of uselessly repeating code. This simplifies things.