Removed null checks
Removed the null checks from the QuestionManagement initialize method. This is mainly because with the null checks, it would interfere with the redesigned locale change hook (416c4e517689e53ab39b486ea90aa69414fc74d7), but also because it's unnecessary. The only reason those checks are there is because there's the possibility of the method being called elsewhere within the app, and as a result the code would be run twice unnecessarily, and I thought including the null check would save cycles. But really, it's not necessary, in any way.