Fixed header size
For some reason, passing 14sp to a TextView object through the setTextSize method is not the same as passing the same value, but stored in Android resources, to the same object. Because it seems after I made that change (70aa010f649114964c3534d29bf465e0e2027ea1), the font size of the header in the digraphs reference screen seems to have doubled. Don't know what the difference is, but because of that I had to decrease the font size.