Altered line breaks in Question Selection
Altered the getQuestionSetDisplay method so line breaks, when the kana question sets are displayed in the Question Selection screen, are prioritized to occur between subsections (so sets of diacritics) rather than in the middle of a subsection. I'm disabling this for any subsection larger than ten items because those are just guaranteed to wrap anyway, and stopping it will only result in improper line breaking, possibly in the middle of diacritics.