Moved final chart setup code
Was occasionally getting a CalledFromWrongThreadException when loading the LogDetailView. Hasn't been happening on all devices, so it took me a while to realize that it wasn't just a weird and coincidental synchronization error. Anyway, I moved the code that was causing the exception to the onPostExecute method, which runs on the main thread. That should solve the error.