Added note for future feature
Added a reminder to try to preserve the current question through a locale change. This'll probably involve storing the key of the current question, and trying to reload it once the new question bank is loaded. But I doubt it's going to be easy, so I just decided to put down a single note so I know to do it later.