Locale change now preserves question
Turns out preserving the current question through a locale change was easier than I thought (7cce3dab5fc8e5d39652ecb176a61b67bfe9dd88). All I needed was a special method in the QuestionBank class for calling up a specific question, and code in QuestionManagement to record and restore that data when needed. The only thing this doesn't do is preserve the previousQuestions record, used to prevent questions repeating too frequently. But I don't really see that as a major issue. It does however know what to do if it can't find the recorded question in the current question bank. In that scenario, it will merely return a 'false,' which tells the calling method that the attempt to load a specific question has failed, in which case, it'll just call the newQuestion method.