Fixed language priority issue
The issue I was having with the wrong language being selected has been resolved by a single line in the build script that does nothing but list every available locale. I broke it down and it seems as if the problem I was having was caused by the app trying to translate into every language. So when it tries to select Dutch, all the dutch locale information from the support libraries is imported along with the default (English) language strings I wrote. But by excluding Dutch from the list, it won't even try to do any Dutch translation, so will skip it and go to the next language in the system priority.