Split up LogDetailView title
In Spanish and Catalan, the title on the log details screen was being truncated as the translations were longer than expected. I tried finding a simple way to just have the title appear on two lines, that ended up being more complex than I thought it would be, and involved expanding the Toolbar class to operate as the new ActionBar. Then I found out I could just set a subtitle on the ActionBar. So by splitting the title at the colon character, I can have the first half remain as the title, and the formatted date in the second half set as the subtitle. Not only does this stop the truncation, it actually looks better, and is more 'fit to purpose.'