Renamed Cherry Blossom colours
The colours for the CherryBlossomTree theme have been renamed to specify that they are part of the CherryBlossomTree theme. As part of this, I had to edit the About screen code, because the java code specifies the specific named colour resource secondaryColor for compatibility reasons, because the theme attribute of colorAccent is incompatible with anything older than Lollipop (API < 21). Anyway, I couldn't specify secondaryColor or any other colour resource because if the theme changes, that resource won't change with it. The only solution I could find was to specify colorAccent in the XML layout document, which I can do on pre-Lollipop, and instead set the text colour to textColorTertiary when it's not operating as a link. Basically, I had to invert the process.