Removed default attribute caching
In a handful of classes, any time I needed access to the default attributes from the theme, or possibly something integrated into the Android frameworks like the selectableItemBackground, I decided to store the values of those attributes in a static variable for use later, so I wouldn't have to keep re-running the code to fetch the data. Thought it would save cycles. But that was written with the assumption that during an app's lifetime, that value would never change. But if one can change the theme of the app within the app, as I'm intending, that's no longer the case. So I decided to remove that little function, and now, when a theme property is needed, it will grab it at the moment, fresh from whatever the current theme is.