Reformatted About screen title
Was getting a weird line wrapping error on the about screen, with each character of the version name ending up on a single line in a single column, to the right of the app name. Of course, this only happens on smaller screens, but nonetheless, I figured I had to fix it. So, I reformatted the layout of app name and version name on the about screen using the FlowLayout class. And that fixed it almost perfectly. Now, if there's any overflow, the version name would just end up below the app name, instead of to the right of it, giving it enough space to display correctly.