Set CherryBlossomTree font
Set the font for CherryBlossomTree to serif. This currently doesn't work for any Japanese characters, because it seems only one version of Noto, Noto Sans, is built-in to Android. And despite the fact that there's an interface to automatically download fonts from Google Fonts, not all fonts seem to be available, even though the documentation clearly states that all fonts are available (https://developers.google.com/fonts/docs/android). This isn't a new problem for me (b53b00dd5a3788c47c5d1bd00f6638ce9ad62c44), but unlike last time, I don't see this problem being resolved by a single line of code. I will have to find a way to integrate the external fonts. Or just accept it as it is.