Fix font on lblResponse
The response label in MainQuiz was using the wrong font. It should've been using the font set by the theme, instead was using the default Android font, Roboto. Turns out, this is because sending null to the setTypeface method wasn't telling the method to just ignore that parameter and leave the font alone, and only set the font style. It was telling the method to set to the system default. I fixed this by fetching the typeface from the object, and passing it back in, in a weird little cycle.