Set fonts on custom views
All the custom view classes that use TextPaint objects have been using the system default fonts, and not the theme fonts. So the obvious solution to this would be to fetch the current font from the theme object. Unfortunately, tried that, didn't work. For some reason, it just returned a null reference. Of course, even if it did work, it would've only worked on Oreo (API 26) and later, because the getFont method is only available for those versions. Regardless, it didn't work, so I had to find an alternative. So I used the really old kludge that I used a long time ago, and thought I'd never have to use again, since I fixed it (9bcb884e425ae0438c9061446a9c8c9abcd6d761). It's remarkably inefficient, because in involves creating a whole new TextView object for the sole purpose of grabbing the font, and then just discarding it. But it'll do for now.