Refined font in theme
Seems I used the wrong attribute when I assigned a font to the CherryBlossomTree theme. Seems 'android:fontFamily' assigns the font to most classes, but I should be using 'fontFamily' because it assigns the font to all the appCompat stuff as well, including the action bar, and the tabs. So I fixed that.