Fixed font kludge
Found a way to fix the font issue. Rather than try to fetch the font from the theme, I'd fetch the name of the font family from the theme (because that's how the fontFamily attribute is stored, as a string), and use that to create a new Typeface object, and use that as the font. Surprisingly simple.