Fixed font attribute again
Funny thing really. When I changed the attribute used to assign a font to the theme (452172dfc001fbec6d59763f2cd21c8a8d62bacd), I didn't fully examine all the effects. Seems both 'fontFamily' and 'android:fontFamily' work on TextView objects, which are the primary objects they need to work on. However, only the 'fontFamily' attribute works on the action bar and the tab layout. Which is why I changed it. But I didn't notice that only the 'android:fontFamily' attribute will work on buttons. So I have to assign both. It's really annoying.