Added fonts to ShibuyaNight
While experimenting with methods to switch between themes, I discovered that some attributes were bleeding through, in a way. When it switched between CherryBlossomTree and ShibuyaNight, the font would remain serifed, even though ShibuyaNight should be sans-serifed. So, by specifying sans-serif fonts in the ShibuyaNight theme, it'll force the correct font to be displayed.