Altered how question type is passed
On the reference screen, question type was being passed through the various tab levels as a string. This was causing errors whenever the strings changed, for instance, during a locale change. So I changed it to pass an integer instead. Said integer being the reference id to the resource string I was previously using. There are still problems being caused by the sub tab references, but this at least solves the first level.