Altered how sub tabs are identified
Continuing from the last commit (95886818edb95c43b1e31cef9f651a9e74b535bd), I altered the way the sub tabs were being identified, so they are passed an integer, rather than a string. Much like the main tabs, it passes the resource reference id of the string it used to pass; unless it's the vocabulary category, in which case it passes the set number. This will probably cause a problem if there are locales in which the number of vocabulary sets in one language are different than the number of vocabulary sets in another language. But for now it's fine.