Altered how vocabulary reference is handled
Created new ReferenceSubsectionVocab class to handle the pages of vocabulary on the reference screen. I mainly wrote it so it could accept a string parameter. because by doing that, I ensure that even if the vocabulary file is radically different between locales, all the index points remain the same. Consistency is what I'm looking for.