Increased consistency of reference page reload
I was looking into how reference pages would reload in changing environments and a problem. To put it simply, if several sets of vocabulary weren't translated, and one was switching between locales, many of the pages on the reference screen end up under the wrong tabs, or just completely blank. I tried several approaches to fixing this, including wiping out the cache of pages from the pager, allowing the cache to be repopulated. But that would just result in the pages getting blanked out permanently. Eventually, I found myself looking through the FragmentPagerAdapter source code, and noticed a line specifying a method called getItemId, which, by default, would just return the position of the page, and is intended to operate as the page's unique id. So, by overriding it to return a different value when certain elements change, It would reload those pages. Surprisingly simple solution to something that was nearly making me pull my hair out. And I got really nice hair, so that would've been bad.