Improved state saving code
The state saving code on the QuestionSelectionPage has been enhanced, so it now recognises more drastic changes to the QuestionSelection options. Continuing my work from the Reference screen (701bdabc4f94b6667cf523aaa0693a5cf60c7d79) I was experimenting to see how the system handles a scenario where the the layout and design of the translated vocabulary file is completely different from the default. Turns out it wouldn't handle it that well, particularly during a locale change. What would happen on the QuestionSelection screen is the items in the new vocabulary file would get selected almost randomly. If they shared a position with a selected item in the previous file, they would be selected in the new file, even if they have a completely different prefId and topic. So by recording the original set of prefIds, I'm able to ensure that the reloaded page would only select items that were selected in the original locale, regardless of their new position. The only issue is, if one is switching back and forth and back and forth between locales, only items that exist in both locales will remain selected. But there's not much I can do about that.