Built theme chooser
Wrote the code allowing a user to select whatever theme they want. It's a bit flawed however. It only works within activities, won't apply to the multitasker or during app startup, and once a theme is selected in options, one has to rotate the screen at least once to activate it. Will have to find ways to resolve this. I added a note reminding me to at least find a way to set the theme's scope app-wide, because that's the one I'm honestly more likely to forget.