Changed manifest theme
Changed the manifest theme to a basic dark theme. It's basically all-black theme from the support libraries that's meant to be expanded upon. This is so, when the app is loading, it's just an all-black screen, with no theme-specific colours. It only affects initial loading, and the multi-tasker. It won't affect any of the activities, or anything in the app. This does mean I no longer need to worry about trying to set the user's theme app-wide.