New ThemeManager class
Created a new class to deal with all theme-related code. Started by moving the getCurrentTheme code for the theme chooser (0c25edd32637043099a62bb376d3703214dad4f7) from OptionsControl to the new class, and renaming it to a more accurate description of getCurrentThemeId. Because it's not passing the theme itself, but the theme's id.