Streamlined theme attribute fetching
Created specialized methods for fetching theme attributes. Because rather than use the obtainStyledAttributes method, passing it an array of one attribute id, and calling a method on top of that, condensing it into a single method makes things much simpler. It will also allow me to redesign this for efficiency, if necessary.