Added support for RikkaApps' Font Provider
The RikkaApps Font Provider (RikkaApps/FontProvider) is an interface to allow the complete Noto CJK fonts, which otherwise have only limited support on Android by default, to be downloaded and accessed in full by any app that supports their API. So I thought I'd give this a shot. Might be the most reliable way to support alternate fonts of Japanese characters. As part of this, I have to abandon support of pre-Lollipop devices (API <21). But given that's only around 10% of the market share, and at least five years old, that's a price worth paying at this stage, in my view.