Rectified font issue
Was having a problem with the Font Provider interface. Without the interface installed, all the latin letters would be serifed; but with it, all the Japanese letters would be serifed; but never both. I found the solution in the Font Provider source code's sample program. Seems there was an alternate 'replace' method I was supposed to be using for serifed fonts that would accept details on the fonts it would replace. At least, I think that's how it works. I feel like this is a documentation failure, but nonetheless, I fixed the issue. (Fixes RikkaApps/FontProvider#9)