Split up the theme fetching method
Split off some of the code in the method used to fetch themes, so there's one method that gets a theme's resource id for a given prefId, and another that just gets the current theme's resource id. This is intended for flexibility, and because I'm working on a new, enhanced, theme chooser, and it'll require this kind of thing.