Added alerts for Font Provider errors
The new Font Provider system has a major flaw: There's an external app the user needs to install, otherwise the fonts won't work properly. And if they don't know to install the app, then the app's fonts will continually be out of whack. So I added a quick alert that checks if the Font Provider app is not installed, and in that case tells the user they should install the app if they want decent fonts. There's also a second alert telling the user if the app is out of date, and that they should update it. I wanted to add two links, one to Google Play, the other to GitHub. But seems there's a limit to how many buttons there can be on an AlertDialog. I think I have a workaround, but I'll have to experiment a bit.