Altered download link
Altered the download link in the 'Font Provider missing alert box' to change depending on whether Google Play is installed. If it is installed, it'll direct the user to download the app from Google Play. Otherwise it'll provide a GitHub link. There does seem to be a few issues with false positives. I might have to refine this a bit more.