Moved theme setting
Moved the theme setting code to the very beginning of the setTheme method. This is because the dialog boxes used to alert users to problems with Font Provider have been in the wrong colour scheme. They should match whatever theme the user set for the rest of the app. This change fixes that.