Added missing fonts message
Found a way to detect if the user downloaded the required fonts through the Font Provider app, and how to open the app at the press of a button. Combined this to present a message to the user if the required fonts are unavailable, and allow them to easily open the Font Provider app to download them.