Fixed region issue on theme preview
The section of the theme chooser designed to preview the theme's font is displaying the wrong kanji. It's displaying the chinese version of the characters, not the japanese version. To fix it, I'm replicating my fix from the kanji-questions branch (b53b00dd5a3788c47c5d1bd00f6638ce9ad62c44).