Fixed border in ThemePreview
With the most recent change to the theme chooser (b927389abaf2f492bc2a33616f376811e0138628), where I removed the top and bottom padding on the ThemePreview object, and let it just float in the centre, turns out I broke the borders. The borders in the ThemePreview class are just there to make it more viewable, because without it, there's a decent chance the background in the preview will just blend into the background of the chooser. But I wrote the code for the border so it builds the border in the padding. This is fine, as long as there is padding. Without the padding, it just gets cut off. So I rewrote the code so it draws the border inside the proper content area, and pushes the theme elements further in. Also changed it so the dimensions of the border are calculated in the onSizeChanged method like everything else, and not the dispatchDraw method like it was before.