Changed colours to be legible
There are a couple areas where the app fetches colours specific to the task at hand, directly from the colours in resources, rather than through the themes. This results in some problems regarding readability, because the colours were originally chosen to be on a light background, and on a dark background, some of them are unreadable. I fixed this by doubling up on all the colour resources. They now all reference either a dark or light state, signifying they're appropriate for either a dark or light theme. To manage this, I created a new method in ThemeManager to check if the theme is a light theme, by fetching the attribute that identifies the theme as such. Then I had to use if-else statements in the locations where the colours are used to split it up so only the colours labelled 'dark' are used in a dark theme, and only colours labelled 'light' are used in a light theme. As part of this, I did have to change the getPercentageColour method to accept a Context object instead of a Resources object, because I kinda need a Context object to get the theme parameters, and I can just get the Resources object from the Context anyway.