Added support for separated answers
Noticed that one of the new vocabulary questions (Tokei) has multiple answers split by a slash (Watch/Clock). Rather than split them up and assign one as an AltAnswer, I decided to instead support both answers in the one attribute, much like is done in the KanjiQuestions class; because they're not synonyms, and I want both answers to appear on the reference screen and multiple choice panel. AltAnswers should be reserved for synonyms. This was done by copying the relevant code from the KanjiQuestion class into the WordQuestion class.