Removed small files
There were four files that were a bit too small. Two were enumerations, one was an interface, and another was a class with only one real method. More than that, they were each used primarily in only one class. So I decided to move the contents of each of those files so they become subclasses of the classes they're primarily accessed from. So I moved the Diacritic enum to the QuestionManagement class, the RomanizationSystem enum to the KanaQuestion class, the OnAnswerListener interface to the AnswerFrame class, and the QuestionRecord class to the QuestionBank class. I also had to make a few changes to the XmlParser and MultipleChoicePad classes because they did access these removed files, but only as a secondary thing. So I think logically I'm sound.