Corrected Font Provider names
In a couple of the strings used in dialog box messages, I forgot to put a space in the name of the Font Provider app. I think this was because the GitHub repository name doesn't have a space in it, and I was looking at that at the time. But the app itself is actually named "Font Provider" with the space. So I fixed that.