Added Font Provider link
Added a link to download the Font Provider app to the Options screen. This is so, even if the user decides to choose the 'ignore' option on the dialog box that suggests they download the app in the first place, there's still a quick and easy way to download the specific Font Provider app. This is actually quite necessary since without it, if the user tries to search for the Font Provider app somewhere else, they might find a different app and think it's the right one when it's not and just get confused when it doesn't work. This fixes that problem.