Added check for Marshmallow
In Android Marshmallow (API 23), one needs to grant permission for the app to access the Font Provider fonts. Thankfully this doesn't apply to any version before or after. One problem I ran into was getting the activity to restart once permission is granted. Because without that, the user will have to browse around until the activity restarts through normal operation before the fonts are applied. Added a note to remind me to fix this in the future.