Suppressing API check
There appears to be a not-insignificant glitch in the NewApi lint inspection. Even though I blocked off all the Marshmallow-specific code behind an API check (1a96abc1d47769f6e32d22281acf7a4576af1185), it's still detecting it as not-matching the correct API. This is really irritating, but it means I need to outright suppress the inspection altogether. I could solve this by restructuring the code, but let's not go that far just yet.