Consolidated variable code
The dialog box building code that directs users to download/update the Font Provider app uses a conditional statement to determine the ideal download location. But there is really only two variables that change between the two conditions, and the rest of the code repeats. So I moved the conditional statement so it sets two variables in advance, and had the variables feed into the method call. This reduced the amount of repeating code.