Consolidated 'remind tomorrow' records
Having two preferences stored for the 'remind tomorrow' preference (one for new installs, one for updates) is pointless. Because it's highly unlikely that both scenarios are going to occur in the same day. If they do a fresh install of the app, it's going to be the newest version, so an update alert just won't happen. And if they need to update, ask to be reminded tomorrow, and then uninstall the Font Provider, I doubt they want to be alerted to undo what they just did. So it makes sense to only have one preference storing this condition.