Fixed font handling for pre-Oreo
Since I did most of my testing on Oreo (API 26), I had no idea about one crucial flaw in the design: The fontFamily attribute doesn't really exist pre-Oreo. This manifested itself in the getThemeFont method, by the getString method returning not the name of the font stored in the theme's fontFamily attribute, like it normally does on Oreo devices, but instead returning the path of some random xml file I don't recognise. And when that's passed to the 'Typeface.create' method, it just ignores the parameter. For some reason, Android Studio doesn't flag it as a problem, which it normally does, so I had to figure this out on my own through some complex debugging. The solution I eventually found was to use the typeface attribute. But when the attribute value is fetched in code, it returns an integer, and not the the actual string value. This is due to how the attribute is handled. It expects one of only four values, and stores this as an integer. Thankfully, this is actually well-documented. So I was able to modify the getThemeFont method to handle the typeface attribute, replacing any integer values with the proper fontFamily name. But because this is specific to certain versions of Android, I also had to write a special getDefaultThemeFont method to handle the switching of the two attributes, and replaced all uses of the getThemeFont method with that. But the weirdest problem I ran into was this: Including both android:typeface and android:fontFamily attributes would work fine, except buttons on pre-Oreo devices would now just ignore both. It would work fine with either one attribute or the other, but not both. So I figured the simplest solution was to split up the theme file into one for pre-Oreo, and one for Oreo and later. And while I was at it, I just had pre-Oreo theme file replace the base theme file that was only being used for pre-Lollipop (API <21) devices anyway, which was rendered pointless since I abandoned pre-Lollipop support a while ago, when I decided to utilize the FontProvider framework (e12fbd79cba9e68effbd361a0967c79f9be4baf4). It appears the basic 'fontFamily' attribute (no 'android:' prefix) is fine though, since I think it's the one utilized by the support libraries. I also have to include the basic 'fontFamily' attribute, because it's required for the FontProvider library to work. However, even though I got this mostly working, the ActionBar and TabLayout objects, as well as the AlertDialog messages, are still not using the correct fonts on pre-Oreo. I'm not sure if there is a fix for that, since I can't find a way to directly change those fonts. However, this is better than before, so I'll leave it for now.