Reorganized Extended Katakana
The extended katakana set has been reorganized into four categories: two for the list set by the Japanese Ministry of Education, Culture, Sports, Science and Technology (or MEXT; and no, I have no idea how that acronym works given that it doesn't even match the Japanese name); one for the list suggested by the ANSI and BSI, both English organizations, so barely relevant; and a final list for those from "the 1974 version of the Hyōjun-shiki formatting". I got these categories from Wikipedia, and I'm not sure if it's the most accurate, but it'll do for now. Because the words 'Extended Katakana' are used in all set names, I decided to set it as a string resource for formatting, much like I did with the normal set names. This required a slight alteration to the XmlParser, to start considering the second string format parameter, and replacing it with the correct value. Also, since I renamed the string resource 'set_11_title' to 'extended_katakana_title', I had to extend that to the translated string resource files. The only major issue that appeared while developing this was in the conflicts I discovered a while back between my sources and Wikipedia (990273a8c2e12ccfbcb28439bd8d0d4b7d99e61c). I had to choose between Wikipedia, and either Learn Japanese To Survive or my Living Language textbook. I think it would've been best to choose LJTS or Living Language, but I just took a careful look at each 'answer' value, and chose the one that wasn't also used for the romanization of another katakana value elsewhere, because I felt that was most likely the correct answer. This resulted in me choosing Wikipedia for all four. I placed the alternates in AltAnswer elements. The end result is, the extended katakana list, which was originally supposed to be an experiment, is finally ready to be released to the world.