Encapsulated QuestionManagement objects
The various QuestionManagement objects were exposed as public objects to the rest of the app. This is not exactly ideal, so I encapsulated all of them so they can't be altered by external classes. I was also going to make the QuestionManagement constructor private, but the XmlParser still needs access to it in the parseXmlFileSetDocument method. So I just made the constructor package-private. It'll do for now.