Modified dialog boxes
The dialog boxes suggesting the user download the Font Provider app had a major problem on Lollipop (API 21 & 22) devices. The buttons were actually overflowing the box, which meant some of them couldn't be accessed. This was unacceptable, so I found a fix to get them to render vertically, like Android does by default on later versions.