Recoloured navigation bars
The navigation bars in each of the themes were too light. This is a problem because the icons are white, and white on a light background is practically invisible. Also, there doesn't appear to be a way to change them in anything earlier than Oreo's second revision (API 27) which introduced the windowLightNavigationBar attribute. So, to simplify things, I decided to just darken the navigation bars. It's the simplest solution. Since I finally tested it from every angle, I think this basically renders the whole dark theme project complete. So this effectively closes #8.