Reorganized Marshmallow code
The code for checking the permissions on Marshmallow (API 23) devices was causing a weird glitch in with the lint inspector which I had to supress (642ec643dfd4d82d25831fa26e34c05cf67b1427). Because it slightly annoyed me, I reorganized the code so it would no longer trigger the glitch. This changes nothing in functionality.