Set app to enable fonts once permission is granted
The code for dealing with permissions on Marshmallow (API 23) had a slight problem. Once permission was granted, there was no simple way to restart the activity so the fonts could be initialized. Instead the fonts would only enable once the user browsed around, and the activities restarted on their own (1a96abc1d47769f6e32d22281acf7a4576af1185). The only way I found to resolve this was by overriding the onRequestPermissionsResult method in each activity. So that's exactly what I just did. To simplify the code a bit, I only did this in activities that used Japanese characters. The Log View, Options and About screens don't use any Japanese fonts, so I did nothing there. I also placed the bulk of the code within a method in the ThemeManager class, and had the overridden onRequestPermissionsResult methods just call that. All the code does is check if the necessary permission has been granted, and will only restart the activity if it has. Otherwise the user could just keep denying, causing it to reload ad infinitum, which is not good.