MainQuiz restarts more frequently
The code to restart the current activity once permissions are granted on Marshmallow (API 23) (e542e7abfe27302f5eda5042d457d22c809a6b11) had a minor flaw. It wouldn't work if one was in an activity like the ReferenceScreen, and returned to the MainQuiz activity. Because, while the fonts would be loaded on the ReferenceScreen, they wouldn't be loaded on the MainQuiz. This is due to the fact that only activities capable of changing anything regarding the available questions, that being the QuestionSelection and OptionsScreen, will restart MainQuiz once one returns to it. To fix this, I tried setting the permissionRequestReturn method to fetch the parent activity, and if it is the MainQuiz activity, to reset it. But that didn't seem to work. After double-checking the documentation, I noticed the getParent method was for activities that are like views or fragments physically within activities, and not activities that are within activities in a navigation sense. And even if it did work, it would still fail if we were in the AboutScreen, because permissionRequestReturn isn't called from there. So I found the only practical solution was to set the MainQuiz activity to restart any time one returns from any activity, no matter what the activity is. It may seem slightly wasteful, but I don't think it'll add enough overhead to make a much of a difference, in terms of end-user experience.