Migrating NumberPreference code
The recent migration of the PreferenceFragment code to AndroidX (67889f3819fb668d86688f3dcfc61d818c10999a) caused some problems. The fact that the NumberPreference class was extending a class from the android.preference package was causing a crash in the OptionsFragment code, because it expects a preference sourced from the AndroidX libraries. But migrating the NumberPreference code was not a simple task, because it seems the DialogPreference class was completely redesigned in its transference from the main Android libraries to AndroidX. In order to manage the change, I had to split it up into two classes, one for the preference itself, and another for the dialog box. I don't see what the point of this is, given the two are inescapably linked. I cannot use the NumberPreference objects with any other dialog, and I can't use NumberPreferenceDialog objects with any other preference. But regardless, this is the new design, so I decided to roll with it. It required overriding the onDisplayPreferenceDialog method in OptionsFragment to get the dialog class to open when the preference object is tapped. I also had to do a bit more than just split up the code across the two classes. Seems the 'persisted value' isn't a thing anymore, so I had the dialog just call the preference value directly from the OptionsControl class, and save it as well, because ironically there doesn't seem to be a way for the preference object to handle it. Had to go through a lot of trial and error to get everything working. But it seemed to work in the end. Happy with the result.